feat: Add `expiresIn` option
This allows users to override the default expiry behavior when using
DynamoDB TTL to expire sessions from the table. This would be useful,
for example, if they wanted to use a browser session scoped cookie as
the session cookie (with no maxAge attribute) but needed a shorter TTL
than the default '1 day from now'.

If this parameter is not provided then there will be no change to the
current behaviour.